requirements_dev.txt
hypothesis==3.8.2
pytest==3.0.7
PyYAML==3.12
